+++ web/src/components/resourceTypes/component/detail/index.js
