+++ web/src/components/workflow/component/add/Form.js
@@ -101,7 +101,6 @@ class AddWorkflow extends React.Component {
       metadata: { name: '', annotations: { description: '' } },
       stages: [],
       currentStage: '',
-      resourcesArr: [],
     };
 
     if (!_.isEmpty(data)) {
