+++ web/src/locale/zh-CN.yaml
@@ -219,6 +219,7 @@ resource:
   operations: 支持操作
   binding: 集成服务绑定
   parameters: 资源参数
+  selectResource: 选择资源
   messages:
     typeRequired: 资源类型必填
     resolverRequired: 资源解析器（镜像）必填
